requirements.txt
coloredlogs==15.0
cycler==0.10.0
decorator==4.4.2
geographiclib==1.50
geopy==2.1.0
humanfriendly==9.1
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5.1
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
tifffile==2021.6.14
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
